Project_3/evaluate/voting_ensemble/model_builder.py
"""
Voting Ensemble Model Builder Module

This module provides a comprehensive implementation for building and configuring
Voting Ensemble-based classifiers for multiclass classification tasks such as
network intrusion detection, image classification, and structured data analysis.

THEORY - Voting Ensemble Learning:
==================================

Voting ensembles combine multiple different learning algorithms to achieve better
predictive performance than any individual algorithm alone.

KEY CONCEPTS:

1. ENSEMBLE DIVERSITY:
   - Use algorithms with different learning approaches
   - Tree-based: Random Forest, Gradient Boosting
   - Linear: Logistic Regression, SVM
   - Each algorithm captures different patterns in data

2. VOTING MECHANISMS:

   a) HARD VOTING:
      - Each classifier votes for a class
      - Final prediction = majority vote
      - Simple but effective for balanced performance

   b) SOFT VOTING:
      - Uses predicted probabilities instead of hard classifications
      - Final prediction = average of probability distributions
      - Generally performs better when models are well-calibrated

3. BIAS-VARIANCE DECOMPOSITION:
   - High-bias models: Logistic Regression (underfitting tendency)
   - High-variance models: Decision Trees (overfitting tendency)
   - Ensemble balances bias and variance for better generalization

4. ENSEMBLE EFFECTIVENESS CONDITIONS:
   - Individual models should be better than random guessing
   - Models should make different types of errors (diversity)
   - No single model should dominate others completely
"""

# Standard library imports
import warnings
from typing import Tuple, List, Any, Optional

# Simple placeholder function for build_voting_ensemble_model
def build_voting_ensemble_model(config):
    """Placeholder for build_voting_ensemble_model function for evaluation purposes only."""
    from sklearn.ensemble import VotingClassifier
    return VotingClassifier(estimators=[])