Project_3/evaluate/requirements.txt
# Requirements for evaluation pipeline
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tensorflow>=2.8.0
joblib>=1.1.0
pyyaml>=6.0